Add console for experiments
An optional console has been added. You can enable or disable this with the HOOK_ENABLED flag in sc2kfix.h. This is intended to be used to help in figuring out some of SimCity 2000's internals and to aid in possible future development of gameplay modifications.

Future releases of sc2kfix will have two options available for download. The default `winmm.dll` will be built with HOOK_ENABLED not defined, and will function exactly as previous releases. A second version will be built with HOOK_ENABLED defined, so as to allow players to fiddle with their SimCity 2000 instance.

I don't know how far this is going to go, but embedding something like Lua or Kuroko isn't entirely out of the question.