Implement multi-threaded music engine
This should eliminate freezes when new music plays, which seemed especially noticeable on Windows 10 and 11. If you're reading this and you like compiling things from source, please give this commit a go. If you run into issues, please try to recreate them with MCI_DEBUG_THREAD enabled and open a GitHub issue with the results.