Move military base hooks out into their own module
I'm working on figuring out how to generate naval bases, so this code is going to get too gnarly to live in the middle of the mischook code. Especially since it might end up being something allowing me to create new types of military bases down the road, which would definitely need to live somewhere else.