Clean up several modules using GAMECALL interfaces
This is a big cleanup to make a number of hooks and associated code use GAMECALL interfaces instead of blasting out raw assembly. It's a lot cleaner, a lot easier to read, and massively more reproducible because it uses more or less plain C/C++ syntax to make calls into SimCity 2000's own code instead of manually pushing arguments onto the stack and calling through registers and using push/retn combinations as a janky jmp instruction.

Going forward we can use the GAMECALL and GAMEJMP interfaces to call into the game engine cleanly without needing to write piles of really nasty assembler scaffolds, and it'll make maintaining the plugin a lot easier.